all: remove redundant type conversion
GitHub-Last-Rev: 6682724eb7cce8e3194650d28cb5eebd78e14700
GitHub-Pull-Request: golang/tools#396
Reviewed-on: https://go-review.googlesource.com/c/tools/+/428980
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Robert Findley <rfindley@google.com>
gopls-CI: kokoro <noreply+kokoro@google.com>

+++ cmd/splitdwarf/splitdwarf.go
@@ -182,7 +182,7 @@ for input_exe need to allow writing.
 		oldsym := symtab.Syms[ii]
 		newsymtab.Syms = append(newsymtab.Syms, oldsym)
 
-		linkeditsyms = append(linkeditsyms, macho.Nlist64{Name: uint32(linkeditstringcur),
+		linkeditsyms = append(linkeditsyms, macho.Nlist64{Name: linkeditstringcur,
 			Type: oldsym.Type, Sect: oldsym.Sect, Desc: oldsym.Desc, Value: oldsym.Value})
 		linkeditstringcur += uint32(len(oldsym.Name)) + 1
 		linkeditstrings = append(linkeditstrings, oldsym.Name)
